+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * New [/holidays](./holidays), a curated list of Bitcoin 'Holidays'
 * Support for different view options on [/fun](./fun)
 * On [/difficulty-history](./difficulty-history), make delta graph honor timespan filtering
+* Proper use of production-ready MemoryStore for session data
 * Support for serving static assets via a configurable CDN
 * Misc fixes for erroneous data display on non-mainnet nodes
 * Switch from fontawesome to bootstrap-icons v1.8.0